multi-word searches
This allows searches like ‘department transport’ to match ‘Department
of Transport’ by splitting the search query into words, and searching
the organisation name for each of the words. Only if the ALL the words
match an organisation name does it appear in the autocomplete.

Thanks to @edwardhorsford for the idea.